# [Performance] Dynamic Batch Tokenizer