## Motivation
- This PR introduces an `AsyncDynamicBatchTokenizer` that enables batching of tokenization requests to improve throughput and reduce latency for SGLang's tokenizer manager.

**Performance Impact:**
- For Qwen3-Embedding-0.6B with 500 input tokens and 1 prompt per request at RPS 500, P99 latency improved from **5335 ms to 463 ms** (~**11.5× faster**) with this PR.

## Context

This PR builds upon the tokenization batching infrastructure introduced in [PR #5141](https://github.com/sgl-project/sglang/pull/5141), which added `enable_tokenizer_batch_encode` for batching multiple texts within a single request.

### Tokenization Batching Options

| Feature | Use Case | When to Use |
|---------|----------|-------------|
| `enable_tokenizer_batch_encode` (PR #5141) | Client sends **batched inputs** in single request | When your client batches multiple texts into one API call |
| `enable_dynamic_batch_tokenizer` (This PR) | Client sends **single prompts** across multiple requests | When your client application sends individual requests |

**Example scenarios:**
- **Use `enable_tokenizer_batch_encode`**: Client sends `{"input": ["text1", "text2", "text3"]}` in one request
- **Use `enable_dynamic_batch_tokenizer`**: Client sends multiple concurrent requests: `{"input": "text1"}`, `{"input": "text2"}`, `{"input": "text3"}`


# Modifications
- 🚀 Dynamic Batching
  - Automatically batches multiple concurrent tokenization requests for efficiency
  - Processes single requests immediately when no other requests are pending
  - Collects additional requests up to `max_batch_size` or `batch_wait_timeout_s` when queue has pending items

- ⚙️ Configurable Parameters
  - `max_batch_size` (default: 32): Maximum number of requests to batch together
  - `batch_wait_timeout_s` (default: 0.002s): Maximum time to wait for additional requests
  - `enable_dynamic_batch_tokenizer`: Feature flag to enable/disable the functionality
  - Usage
    ```
    --enable-dynamic-batch-tokenizer \
    --dynamic-batch-tokenizer-batch-size 32 \
    --dynamic-batch-tokenizer-batch-timeout 0.002
    ```
- 🔄 Async Processing
  - Non-blocking tokenization using `asyncio` and `ThreadPoolExecutor`
  - Maintains event loop responsiveness while handling blocking tokenizer calls
  - Scales efficiently with concurrent requests


# Benchmarking and Profiling
- Model Qwen3-Embedding-0.6B 
- Input Token Length = 500
- GPU Type = H100
- Traffic distribution = Poisson

- Baseline numbers
```
(li-sglang) jobuser [ /shared/user/repos/li-sglang ]$ python -m sglang.launch_server --model-path /shared/public/sharing/suramach/Qwen3-Embedding-0.6B --port 30000 --host 0.0.0.0 --enable-metrics --disable-radix-cache --disable-cuda-graph  --max-prefill-tokens 131072 --is-embedding
```

```

Overall Summary for RPS 500, Duration 120s, Item Count 1: 
-- | -- | --
Test duration:         120 seconds  
Server type:           HTTP |   |  
HTTP mode:             EMBEDDINGS |   |  
Target RPS:            500 |   |  
Item count:            1 |   |  
Distribution:          POISSON |   |  
Unique requests generated: 100 |   |  
Total requests sent:   60000 |   |  
Successful requests:   60000 |   |  
Failed requests:       0 |   |  
Time to send all requests: 162.92 seconds |   |  
Time for all requests to complete: 168.23 seconds |   |  
Average response time: 2408.00 ms |   |  
P50 response time:     2561.12 ms |   |  
P90 response time:     4234.68 ms |   |  
P99 response time:     5335.20 ms |   |  
```

- With this PR
```
(li-sglang) jobuser [ /shared/user/repos/li-sglang ]$ python -m sglang.launch_server --model-path /shared/public/sharing/suramach/Qwen3-Embedding-0.6B --port 30000 --host 0.0.0.0 --enable-metrics --disable-radix-cache --disable-cuda-graph  --max-prefill-tokens 131072 --is-embedding --enable-dynamic-batch-tokenizer
```

```
Overall Summary for RPS 500, Duration 120s, Item Count 1: |   |  
-- | -- | --
Test duration:         120 seconds |   |  
Server type:           HTTP |   |  
HTTP mode:             EMBEDDINGS |   |  
Target RPS:            500 |   |  
Item count:            1 |   |  
Distribution:          POISSON |   |  
Unique requests generated: 100 |   |  
Total requests sent:   60000 |   |  
Successful requests:   60000 |   |  
Failed requests:       0 |   |  
Time to send all requests: 164.05 seconds |   |  
Time for all requests to complete: 164.87 seconds |   |  
Average response time: 163.23 ms |   |  
P50 response time:     114.94 ms |   |  
P90 response time:     356.91 ms |   |  
P99 response time:     462.65 ms |   |  


```


## Checklist

- [ ] Format your code according to the [Format code with pre-commit](https://docs.sglang.ai/developer_guide/contribution_guide.html#format-code-with-pre-commit).
- [ ] Add unit tests according to the [Run and add unit tests](https://docs.sglang.ai/developer_guide/contribution_guide.html#run-and-add-unit-tests).
- [ ] Update documentation according to [Write documentations](https://docs.sglang.ai/developer_guide/contribution_guide.html#write-documentations).
- [ ] Provide accuracy and speed benchmark results according to [Test the accuracy](https://docs.sglang.ai/developer_guide/contribution_guide.html#test-the-accuracy) and [Benchmark the speed](https://docs.sglang.ai/developer_guide/contribution_guide.html#benchmark-the-speed).